refactor(provider): extract provider resolution into resolve.rs (dirge-4y4l stage 5)
Split provider/mod.rs (2217 LOC) by moving the pure
provider-resolution surface into a new provider::resolve submodule:
ProviderKind, model-default lookups, parse_provider, ProviderInfo +
resolve_provider_info, custom/plugin endpoint validation, env-var
autodetect, and resolve_api_key. mod.rs re-exports via `pub use
resolve::*` so the public path (provider::parse_provider etc.) is
unchanged. Test-referenced helpers bumped to pub(crate).

Behavior-preserving. mod.rs 2217->1738. Builds clean -D warnings on
default + windows-default (no-plugin); 2374 tests green.
